// src/data/enum/revalidate_tags.enum.ts
export const enum RevalidateTagsEnum {
  POSTS = "posts",
  POST = "post",
  DESTINATION = "DESTINATION",
  AREAS = "AREAS",
  SALETRANSACTIONTYPES = "SALETRANSACTIONTYPES",
  OWNERTYPES = "OWNERTYPES",
  BUILDINGPROPERTYTYPES = "BUILDINGPROPERTYTYPES",
  BUILDINGOWNERSHIPTYPES = "BUILDINGOWNERSHIPTYPES",
  LANDPROPERTYTYPES = "LANDPROPERTYTYPES",
  LANDOWNERSHIPTYPES = "LANDOWNERSHIPTYPES",
  PASSPORTBYID = "passportByid",
  PASSPORTGETALL = "passportGetAll",
  PASSPORTFILETYPES = "passportFileTypes",
}